local: in progress hackery for trying to bring my sgi indy up
this controller driver doesn't handle errors well.

I have it not panicing, but now I can see it stick in a timeout
that leads to an ABORT, then but reset, then select, then the DMA
transfer hangs, and I'm not yet sure why.

I can repro this by doing filesystem IO with this diff, just
cp'ing /sbin/* into the FFS on the zuluscsi SD on /dev/sd0a .

This is happening in a loop when things out to lunch:

[  86.7709586] sd0(wdsc0:0:1:0): wdsc0: timed out; asr=0x20 [acb 0x97ec4fa8 (flags 0x41, dleft 20)], <state 5, nexus 0x97ec4fa8, resid 20, msg(q 0,o 0)>
[  86.9268976] sd0(wdsc0:0:1:0): ABORT in timeout: csr=0xff, asr=0x20
[  87.0011412] sd0(wdsc0:0:1:0): sending ABORT command
[  87.0597051] sd0(wdsc0:0:1:0): Resetting bus
[  87.1099010] available sync periods: 50 75 100 125 150 175 200
[  87.1789261] available fast sync periods: 25 37 50
[  87.2353922] wd33c93_sched(1,0)
[  87.2719980] wd33c93_go(1:0)
[  87.3054569] wd33c93_go dmago:1(tcnt=20) dmaok=1x
[  87.3608921] wd33c93_selectbus 1: Selection Complete
[  87.4194486] wd33c93_setsync: sync reg = 0xac
[  87.4706861] go[0x1a] next[a=00,c=1a]: wd33c93_xfout {6} 03 00 00 00 20 00 00 00 80 00
[  87.5648055] wd33c93_xfout done: 0 bytes remaining (wait:50000)
[  87.6348702] go[0x19] next[a=80,c=19]: DMA xfer: 1(0x97f111dc:20)
[  87.7070405] > done i=1 stat=ff

i=1 stat=ff means SBIC_STATE_RUNNING, so the transfer itself hasn't completed,
and then it will eventually continue to time out.

So, I need to go continue to figure out what's broken in this loop and
figure out how to get the drive back.

(But at least it's not panicing now.)