wd33c93: split wd33c93_timeout() into timeout and callback
The code actually directly calls the same function for both
processing a timeout (and thus sending an abort, bus reset, etc)
and the callout running.

However, the timeout code doesn't actually complete an acb.
So if an acb times out, it seems to just .. be timed out.
Forever.

So in preparation for adding support for completing an acb with
timeout, split wd33c93_timeout() into timeout and callout.

This should be a no-op.